variation ordering, and variation ui changes for dan,
UI changes for variations: have removed the quick edit functionality as
we want to have the default view as the quick edit mode this will make
things easier and more clear for the users. This is unfinished and will
need lots of testing to make sure none of the other tables using
wpsc_product_row() have not been broken.

git-svn-id: https://wp-e-commerce.googlecode.com/svn/trunk@1183 7bdd534d-2f5a-8b57-a18c-7e4837153298