Fix: Category title is replaced with the first product's title, and pagination does not work correctly on Category page.
The first bug is a subtle one. To reproduce, the theme usually has something
like this:
<a href="xxx" title="<?php the_title();?>"><?php the_title(); ?></a>

The first "the_title()" call is replaced correctly, but not the second.

To fix this, we direcly modify $wp_query->post->post_title.

The next bug is about pagination. The "posts_per_page" of the main
query falls back to using WordPress core blog posts per page option.

Fixes issue 1159.

git-svn-id: https://wp-e-commerce.googlecode.com/svn/branches/3.8@2145 7bdd534d-2f5a-8b57-a18c-7e4837153298