Remove 'Delete' sale link from Sales Log page.
Once an e-Commerce site is live, you really don't want to be deleting Sales records. There are legal requirements in most countries to maintain sales records for auditing.
This will disable /hide the 'Delete' button from the sales log page unless the current user is an Admin (Level_10). This allows deletion when developing a site, or under exceptional circumstances for example.

Signed-off-by: Andrew <macbook-131@macbook-131s-MacBook-Pro.local>