Introduce some method parity w/ other new classes
I like having the ability in WPSC_Purchase_Log to filter the entire data array, or each property.  I also like having the ability to easily get_data() and have the whole row available.  When we continue to refactor areas in core, we'll find that many of the functions that we'll use WPSC_Country in will require the entire array of data to be available.