First pass at replacing live/livequery/delegate/on with on()
One file at a time, as this will be quite extensive.  First pass at wp-e-commerce.js

Concerns: There are some places with .livequery() that are not tied to a specific event.  These should be extensively tested.  I don't honestly know why they exist in the manner they do, or how they might break, but they seemed to be okay in my testing.