Fixing Product table layout when no products
This should fix a minor css issue when you don't have any products in the Products table. The Table does not stretch all the way out, there is quite a bit of white space after the date column where the table does not reach. There should be no issues effected by the CSS used to fix the problem!