Introduce changes to fix oft-complained of admin collapse bug.
We haven't found a good reason that this occurs.  Disabling all enqueued JS files from WPEC does not fix it.  In common.js (in WP), the cached body selector simply doesn't get updated when classes are added and removed. When WPEC is disabled, this is not an issue.  Rather than shake our fists at the skies, we're binding a function to an event to handle the collapsing.

This means increasing the scope of the admin.js for WPEC and loading it everywhere.  We may revert this down the line, but for now, it should be fine.

Fixes #2028.  Notably, this will further break #382.