Removed Price tag from Products price amount
Having the "Price:" tag is unnecessary and takes up real estate.
Additionally it requires extra translation.  A price with the appropriate
currency is clear for all users what the item costs without needing a label.

This only affects regularly priced items as sale items are designated with
their own label.