Cart notifications (Including REST API improvements) (#2295)
* Fixes debug notices related to 'sizes' not being set on attach metadata

The 'sizes' parameter in the $metadata array passed through the
'wp_generate_attachment_metadata' is not guaranteed to be there, and
therefore can cause the following notices/warnings:

- PHP Notice:  Undefined index: sizes in
.../wp-e-commerce/wpsc-components/theme-engine-v2/helpers/template-engine.php
on line 481

- PHP Warning:  array_key_exists() expects parameter 2 to be array, null
given in
.../wp-e-commerce/wpsc-components/theme-engine-v2/helpers/template-engine.php
on line 481

* Update WPSC_Template_Engine to handle registering enqueued scripts.

Because of the output buffering, there are many cases where naked calls
to wp_enqueue_script will not work as it happens to early. This allows a
CRUD interface for dealing with those scripts.

* Update js.php to use wpec versions of wp_enqueue_script/localize_script

As per the previous commit, we need a way to enqueue scripts before
they are registered. This allows us to do so as well as provide an
interface for localizing variables within the WPSC namespace.

Also includes a bit of cleanup in copy-billing.js

* First step towards Cart Notifications. Script is enqueued.

* Make sure localized vars don't  break the WPSC namespace. (and our code)

Serializes, ensures sane length and sanitizes.

* Allow third-party devs to set in_footer to false

* s/serialize/maybe_serialize and remove string length limitation for localized vars

* Remove redundant action as wpsc_enqueue_script handles deferring script enqueueing

* Introduce WPSC_Template_Engine::register_default_script_localization

We should set all configuration for WPSC_Template_Engine::$core_scripts
from within WPSC_Template_Engine. Defers until get_core_scripts_data is
called during wp_enqueue_script.

* Update remaining scripts to use wpsc_enqueue_script to enqueue and localize

* Only enqueue fancy notifications if option is set.

* Add css tasks to gruntfile:

"grunt-combine-media-queries"
"grunt-contrib-cssmin"
"grunt-contrib-sass"

* Scaffold underscore template for modal

Allow specifying template to be added to wp_footer via callbacks
with the data passed to wpsc_localize_script.

* Ignore fontawesome file, as it is minified elsewhere

* Run `grunt css` and switch enqueues to use minified if applicable

Will use .min.css files if SCRIPT_DEBUG is not enabled, and
wpsc_use_minified_styles hasn't been filtered to false.

* Fix logic so that filter (wpsc_use_minified_styles) language is correct.

* Include grunt uglify, and create minfied JS resources

Also update watch JS task to only watch a subset of files (for now)

* First pass at some modal styling

* Continue styling modal

* Add 3rd arg to wpsc_product_no_thumbnail_image to allow returning value

* Remove unused strings var

* Update to use backbone and move markup to template files

Also, now working with live data

* Fix modal height max width and update to populate items when clicking add to cart buttons

* Ignore my notes file

* Implement browserify for creating modular JS files.

Specifically for the cart-notifications JS.

* Add a product ID specific class to the widget cart items.

In the future, changes in the cart-notifications modal should also occur
in the cart widget. This paves the way for identifying other cart items
on the page.

* Add a data-id data attribute to the add-to-cart forms.

Allows easy identification of the product ID when dealing with the
form/cart in JS.

* Hook in uglify to watch task

* Major update to cart-notification functionality.

- Tons of improvements. Cart does not actually set the cart items for
  real (yet), but most functionality is in place for interacting with
the cart.
- Modularize cart notifications JS, and use browserify to compile them
  together.
- Move Cart Notification functionality to dedicated class.

* Move implementation out of FloatLabel.js

* Replace _wpsc_enqueue_float_label_scripts with wpsc_enqueue_script('wpsc-checkout' )

* Replace _wpsc_enqueue_product_scripts with wpsc_enqueue_script( 'wpsc-products' )

* Add minified versions of scripts & floatlabel file to allow grunt to minify

* Re-grunt the css files

* Begin adding ajax syncing for cart updates

* Move sync event from collection to model

* Make baseroute generic and test quantity

* Remove nonexistent _wpsc_action_enqueue_shipping_billing_scripts

* show_in_rest for taxonomy registration, and move API classes to dir

* Add rest coupons class

* Properly name the reports controller.

* Default the theme engine to be off if < 4.7.0.
With our checkout process depending on the REST API infrastructure and content endpoints by default, it makes sense to turn this off.
It would be a two-line function to disable fancy notifications and enable the 2.0 theme engine if you knew that was the only reason it wasn't turned on.

* Stub out cart api endpoints

* Stub out functionality for adding item to cart

* Initial pass at documentation improvements in checkout controller.

* Fix namespace, and missing status var

* Need to return $parameters

* Fix error messages missing post title

* Update the since tags

* Improvements to api implemnentation.
Use wpsc/v1 namespace for cart, making cart the rest_base
Ignore core API controllers in bootstrap, as they should be bootstrapped by core.
Stub out schema for controller.

* Set namespace for all controllers and follow register_routes() method conventions.

* /add/<id> should accept PUT since there is an ID

* Update the delete method/endpoint to work

* localise the baseroute and the wp-api nonce

* Update sync method for cart. still needs work

* Update compiled script

* small tweaks/cleanup

* prepare_cart_item_for_js should include nonces for adding/deleting

* Add a remove_item_by_id method to WPSC_Cart for easier removal

* Update to make adding/removing from cart all work properly

* Add minified version of select-to-autocomplete

* Rework grunt to properly minify jquery.* files

* Add edit_item_by_id and get_cart_item_by_id utility methods to WPSC_Cart

* When incr/decr the quantity values, need to trigger "change"

* Add ability to edit cart quantity, and add cart status tab/button

* Revamp WPEC enqueueing system to work in admin as well. Fixes data localization issues.

* Move enqueued scripts to aptly-named method